change path definition rule
Before, loss information and optimzier infomation is both on the same directory name. But here I split them and rightnow is the form of lt_info+loss_name / optim_name. Which makes more sensce because under the first directory collecting all untiary with the same loss, and easier to compare them.